update to ffmpeg 6.0:
from https://github.com/BtbN/FFmpeg-Builds.git

built with commands:

./makeimage.sh win64 gpl-shared 6.0
./build.sh win64 gpl-shared 6.0

./makeimage.sh win32 gpl-shared 6.0
./build.sh win32 gpl-shared 6.0

changed zlib commit to 09155eaa (v1.3)

+++ include/libavutil/bswap.h
@@ -40,6 +40,8 @@
 #   include "arm/bswap.h"
 #elif ARCH_AVR32
 #   include "avr32/bswap.h"
+#elif ARCH_RISCV
+#   include "riscv/bswap.h"
 #elif ARCH_SH4
 #   include "sh4/bswap.h"
 #elif ARCH_X86
@@ -72,7 +74,7 @@ static av_always_inline av_const uint32_t av_bswap32(uint32_t x)
 #ifndef av_bswap64
 static inline uint64_t av_const av_bswap64(uint64_t x)
 {
-    return (uint64_t)av_bswap32((uint32_t) x) << 32 | av_bswap32(x >> 32);
+    return (uint64_t)av_bswap32(x) << 32 | av_bswap32(x >> 32);
 }
 #endif
